Extract _safe as a member variable
This is the first of a series of changes intended to break down the
confusion around default request options and what method takes what
options. _safe is only accepted by the request methods that create HTTP
requests. Its function is to add concurrency control headers like
If-Match and If-Not-Match to requests. Right now it defaults to
false (i.e. off), but I would like to change that on the next major
release.

_safe is now an explicit member variable in KintoClientBase, Bucket, and
Collection. KintoClientBase provides _safe to Bucket; Bucket provides
_safe to Collection. Each method that passes _safe now provides it
explicitly, using its default while allowing user code to override it
for a given method call.