Extract _retry as a member variable
_retry is used by HTTP.request to indicate that a request should be
retried if we get back a Retry-After header. Besides HTTP.request, it is
accepted by execute() and paginatedList().

While we're here, let's formalize the concept of a request as having
only four attributes: method, path, headers, and body. retry isn't one
of those, so let's separate it out into an "options" parameter.