collection_test: Fix test
defaultReqOptions isn't respected at this point in the code. The
listRecords call only merges with the collection's default headers; by
construction of the test, these headers already include Foo and Baz, so
this test was kind of meaningless.

If we intend for client default headers to be taken into account, we
might want to pass them through the bucket at construction time (right
now they aren't unless you call client.bucket). If we intend
for *changes* to those headers to be taken into account, we'll probably
need to work a bit harder.