paginatedList: clean up generated request
These generated requests are for list requests, which means they
shouldn't have a body or a method (it should always be GET). The path is
provided by paginatedList. The only remaining request attribute present
is headers. Use it explicitly rather than unpacking options, which could
have other stuff in it (like `retry`).