Refactor options so we can work with Kinto 7.0.0 (#180)
* Collection: explicitly document options

* Bucket: remove `bucket` from _bucketOptions

This is only used/needed in the call to batch(). It shouldn't be present
in the other places _bucketOptions is used.

* Bucket: remove `batch` from _bucketOptions

This option is only recognized in the Collection constructor, and not in
the other places that use _bucketOptions.

* base: JSDoc a bunch of stuff

Try to firm up the interfaces a bit here rather than just slopping
objects around.

* base: Remove batch from _getRequestOptions

No caller uses this; only batch() cares about this field and it
overrides it explicitly.

* bucket: permissions and data shouldn't come from reqOptions

That would imply that the bucket has "default" permissions and data for
all requests, which is nonsensical.

* bucket: needless spread

* Extract _safe as a member variable

This is the first of a series of changes intended to break down the
confusion around default request options and what method takes what
options. _safe is only accepted by the request methods that create HTTP
requests. Its function is to add concurrency control headers like
If-Match and If-Not-Match to requests. Right now it defaults to
false (i.e. off), but I would like to change that on the next major
release.

_safe is now an explicit member variable in KintoClientBase, Bucket, and
Collection. KintoClientBase provides _safe to Bucket; Bucket provides
_safe to Collection. Each method that passes _safe now provides it
explicitly, using its default while allowing user code to override it
for a given method call.

* Extract _headers as a member variable

This is essentially the same change as _safe, except for
_headers. `headers` can be provided to any request, which means both the
requests methods as well as some methods that don't need to call a
requests method to construct the HTTP request they want to send.

* execute: Fix up defaults

`stringify` was only true if no options were passed at all; actually, we
want it to be true as long as no `stringify` option was passed.

* Extract _retry as a member variable

_retry is used by HTTP.request to indicate that a request should be
retried if we get back a Retry-After header. Besides HTTP.request, it is
accepted by execute() and paginatedList().

While we're here, let's formalize the concept of a request as having
only four attributes: method, path, headers, and body. retry isn't one
of those, so let's separate it out into an "options" parameter.

* collection_test: Fix test

defaultReqOptions isn't respected at this point in the code. The
listRecords call only merges with the collection's default headers; by
construction of the test, these headers already include Foo and Baz, so
this test was kind of meaningless.

If we intend for client default headers to be taken into account, we
might want to pass them through the bucket at construction time (right
now they aren't unless you call client.bucket). If we intend
for *changes* to those headers to be taken into account, we'll probably
need to work a bit harder.

* Get rid of needless and vague "options"

We now explicitly track, as member variables, all the "defaults" which
we use in lots of methods -- _safe, _retry, and _headers. These
"options" fields no longer have anything to contain, so get rid of
them.

* Failing test

* paginatedList: clean up generated request

These generated requests are for list requests, which means they
shouldn't have a body or a method (it should always be GET). The path is
provided by paginatedList. The only remaining request attribute present
is headers. Use it explicitly rather than unpacking options, which could
have other stuff in it (like `retry`).

* integration_test: Test that paginated lists work in batches

This didn't work previously because `retry`, `safe`, `batch`, and others
were sloppily finding their way into all requests, where they would get
rejected by the Kinto server for not conforming with the request body
format it knew. Now that all the options have been cleaned up, this is a
piece of cake.

* utils: Remove getOptionWithDefault

We can do this as a one-liner in the necessary functions.

* ES6 style (thanks @n1k0)

* Flag private methods as @private (thanks @n1k0)

* We don't need these headers here (thanks @n1k0)

* Update documentation regarding request options

* ES6 object members (thanks @leplatrem)

* Always pass patch, last_modified, and aggregate

This required updating some tests, but that seems fine.

* execute: Limit requests to batch-request-safe fields

* Out-format the formatter (thanks @n1k0 for the tip)