base: Remove needless fetchServerSettings from execute()
`fetchServerSettings` calls `fetchServerInfo`, which causes it to fetch
and store the server hello response for future calls to
`fetchServerInfo`. However, other calls to `fetchServerInfo` would also
fetch and store the server hello response if necessary, and the stored
`serverInfo` field is only accessible through the `fetchServerInfo`
method. So the call to `fetchServerSettings` here is meaningless and we
can get rid of it.

paginatedList no longer calls `fetchServerSettings` (because it has no
need to), so update the tests.