Cleanup serverInfo a little bit (#186)
* base: Extract _getHello() and update fetchUser

Unlike the other parts of the "hello" response, fetchUser is not static,
but depends on the Authorization header present in the options, so it
shouldn't be memoized.

* base: Remove `headers` option for fetchServerInfo

We do this explicitly to prevent someone from accidentally sneaking in
authorization information for a response that then somehow gets
memoized.

While we're here, fix up ESDoc for all server metadata methods.

* base: Remove needless fetchServerSettings from execute()

`fetchServerSettings` calls `fetchServerInfo`, which causes it to fetch
and store the server hello response for future calls to
`fetchServerInfo`. However, other calls to `fetchServerInfo` would also
fetch and store the server hello response if necessary, and the stored
`serverInfo` field is only accessible through the `fetchServerInfo`
method. So the call to `fetchServerSettings` here is meaningless and we
can get rid of it.

paginatedList no longer calls `fetchServerSettings` (because it has no
need to), so update the tests.

* base: Pass `retry` to `fetchServerSettings`

This option is recognized, so we may as well pass it here too.

* Prettify code