Cleanup error handling (#256)
* Introduce NetworkTimeoutError for network timeouts

This requires some slight modifications to Babel config to work.

* Don't catch exceptions from formatResponse

Since formatResponse occasionally throws exceptions, wrapping it in a
catch-rethrow block doesn't make sense. I believe this logic is meant
to catch unparseable non-JSON responses, not exceptions thrown by
formatResponse.

* Don't call formatResponse when no response body

This is a small step towards making formatResponse only responsible
for handling error cases.

* Rename formatResponse to throwServerResponse

Invoke it only when the status code indicates that the request
failed. Handle normal responses in processResponse.

* Introduce UnparseableResponseError

Also, include the response body in the error message when this
happens. Mostly these are Invalid Gateway errors so they shouldn't add
undue strain on logs or anything.

* Check for errors even when the body is empty

This probably never happens in practice, but it feels more correct to me.

* Change throwServerResponse into an exception class

* Refactor ServerResponse code according to my personal taste

This is longer but it covers more cases.

* Update tests and add another one to cover missing-body cases