Use transform-runtime to request regenerator without depending on it
transform-runtime is meant to allow us to access the regenerator
polyfill "seamlessly, without global pollution". This means we don't
need to import it explicitly, and we don't need to make any efforts to
import it only once.

This replaces babel-polyfill, which is meant for applications, not
libraries.

Applications are responsible for providing the polyfill.

We use the "polyfill: false" option to disable conversion of "builtin"
JS methods such as Array.from and Object.create, which we could also
rewrite except for
https://github.com/loganfsmyth/babel-plugin-transform-builtin-extend/issues/15.

This is meant to fix #259 and Kinto/kinto#1379.

Refs #159 (cc: @screendriver).

Refs Kinto/kinto-admin#456 and Kinto/kinto-admin#446.