Use sinon.createSandbox instead of sinon.sandbox.create
This seems to be the first of many consequences from
https://github.com/sinonjs/sinon/pull/1586 and
https://github.com/sinonjs/sinon/pull/1700. I'm not sure what the
final result of all of this will be -- we might not even need this
sandbox any more, because a default one is created for us -- but at
some point in the future, we'll probably have to rewrite everything
that uses a stub(), and I prefer to wait until then to make any
drastic changes. This just silences the warnings until then.