Bump to prettier 1.12.0 (#280)
* chore(package): update prettier to version 1.12.0

* Reformat with prettier 1.12.0

* Pin to prettier 1.12.0

This may seem a bit strange, but in a real sense, the build only works
with this version, so we have an exact dependency.

Hopefully this will cause greenkeeper to open PRs to bump it rather
than issues that it breaks the build. Similarly, hopefully this will
prevent me from getting emails from Travis that the build randomly failed.