Restore use of versions.txt when installing master (#317)
* Restore use of versions.txt when installing master

This got dropped soon after it got added, but it seems like this is
why we fetch versions.txt. It's my hope that putting this back in will
cause tests to pass (right now they're failing because kinto relies on
jsonschema==3.0.0a3 and we're installing 2.6.0).

* Bump Python to 3.6

f-strings are not available in Python 3.5, and Travis doesn't yet
support 3.7.